+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -330,7 +330,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVA81EY2Q9BQZRY0TE15KSNW"
+        "01KVAG9AMJ8RVKYN57V0SJDFD0"
     )
     r"""The unique identifier of the data source"""