+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -137,7 +137,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-17T07:33:46.419Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-17T09:58:08.470Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")