+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVA81EXZSHFY9JKH2R10T28K"
+        "01KVAG9AMEQXVVQFMEW0JDXCCD"
     )
     r"""The unique identifier of the data source"""