+++ packages/orq-rc/src/orq_ai_sdk/models/post_v2_feedback_evaluationop.py
@@ -55,7 +55,7 @@ class RequestBody3(BaseModel):
 
     source: Optional[PostV2FeedbackEvaluationRequestBodyFeedbackSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-06-17T07:34:27.226Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-06-17T09:58:33.604Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -121,7 +121,7 @@ class PostV2FeedbackEvaluationRequestBody2(BaseModel):
 
     source: Optional[PostV2FeedbackEvaluationRequestBodySource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-06-17T07:34:27.225Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-06-17T09:58:33.603Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -187,7 +187,7 @@ class PostV2FeedbackEvaluationRequestBody1(BaseModel):
 
     source: Optional[RequestBodySource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-06-17T07:34:27.224Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-06-17T09:58:33.602Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")