+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVA81EY1EC86G7YNBVZ3N198"
+        "01KVAG9AMHV4KXPTBJZCNRV0GB"
     )
     r"""The unique identifier of the data source"""