+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVA81EY5QB95CN85HWV2JVNK"
+        "01KVAG9AMP45H0RDQ5KSMB0EA9"
     )
     r"""The unique identifier of the data source"""